refactor: refactor to promise
