+++ libs/bite-tribe/api/src/lib/restaurant-api/restaurant-api.service.ts
@@ -1,78 +1,17 @@
 import { inject, Injectable } from '@angular/core';
 import { AuthService } from 'ta-firestore';
-import { EMPTY, from, Observable, skipWhile, switchMap } from 'rxjs';
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
 import type { Link, Restaurant } from 'model';
 import { MENU_COLLECTION } from '../menu-api.service';
-import { BITE_COLLECTION } from '../utils/constants';
-
-const RESTAURANT_COLLECTION = 'restaurants';
+import { BITE_COLLECTION, RESTAURANT_COLLECTION } from '../utils/constants';
+import { getRestaurantById } from './utils/load-restaurant-by-id';
 
 @Injectable({ providedIn: 'root' })
 export class RestaurantApiService {
   private readonly authService = inject(AuthService);
 
-  loadRestaurant(restaurantId: string): Observable<Restaurant | undefined> {
-    return this.authService.isLoggedIn$.pipe(
-      skipWhile((isLoggedIn) => !isLoggedIn),
-      switchMap(() => {
-        if (restaurantId) {
-          return from(this.getRestaurantById(restaurantId));
-        }
-
-        return EMPTY;
-      }),
-    );
-  }
-
-  async getRestaurantById(
-    restaurantId: string,
-  ): Promise<Restaurant | undefined> {
-    try {
-      // First try to get by ID
-      const doc = await FirebaseFirestore.getDocument({
-        reference: `${RESTAURANT_COLLECTION}/${restaurantId}`,
-      });
-
-      if (doc.snapshot.data) {
-        const data = doc.snapshot.data;
-        return {
-          id: data?.['id'] || restaurantId,
-          ...data,
-        } as Restaurant;
-      }
-
-      // If no restaurant found by ID, try to find by name
-      const queryResult = await FirebaseFirestore.getCollection({
-        reference: RESTAURANT_COLLECTION,
-        queryConstraints: [
-          {
-            type: 'limit',
-            limit: 10,
-          },
-        ],
-      });
-
-      const restaurantName = decodeURIComponent(restaurantId);
-      const matchingRestaurant = queryResult.snapshots?.find(
-        (snapshot) =>
-          snapshot.data?.['name']?.toLowerCase() ===
-          restaurantName.toLowerCase(),
-      );
-
-      if (matchingRestaurant) {
-        const data = matchingRestaurant.data;
-        return {
-          id: data?.['id'] || matchingRestaurant.id,
-          ...data,
-        } as Restaurant;
-      }
-
-      return undefined;
-    } catch (error) {
-      console.error('Error fetching restaurant:', error);
-      return undefined;
-    }
+  loadRestaurantById(restaurantId: string): Promise<Restaurant | undefined> {
+    return getRestaurantById(restaurantId);
   }
 
   async saveNewRestaurant(restaurant: Restaurant): Promise<void> {
