+++ libs/bite-tribe/api/src/lib/utils/constants.ts
@@ -1,3 +1,5 @@
 export const BITE_COLLECTION = 'bites';
 export const BUCKETLIST_COLLECTION = 'bucketlists';
+export const RESTAURANT_COLLECTION = 'restaurants';
+
 export const DEFAULT_SEARCH_RADIUS_IN_M = 15 * 1000;
